Adjust display positions for card section
The text under each card was not aligning with the image above.
Upon further investigation the block element of image was preventing
the image from being center aligned. Flex was implemented for each card
to solve this issue.

When shortening the width of the browser window the cards were pushed
off screen. I changed the behavior to have the cards stack with
flex wrap.